Implemented command "template-show". Renamed command "pprint" to "template-pprint".
`template-show` just prints the template file as stored to stdout, which means any comments or extra information will be visible.
`template-pprint` passes the template file through the parser and rebuilds it to text reflecting only what would be built case the "run" command was called. So any comments and elements that don't pertain to a directory tree will not be printed.